requirements.txt
absl-py==2.1.0
asn1crypto==1.5.1
attrs==23.2.0
branca==0.7.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
cryptography==42.0.5
filelock==3.13.3
fiona==1.9.6
folium==0.16.0
geopandas==0.14.3
idna==3.6
immutabledict==4.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
networkx==3.2.1
numpy==1.26.4
openrouteservice==2.3.3
ortools==9.9.3963
osmnx==1.9.1
packaging==24.0
pandas==2.2.1
platformdirs==3.11.0
protobuf==5.26.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
shapely==2.0.3
six==1.16.0
snowflake-connector-python==3.7.1
sortedcontainers==2.4.0
tomlkit==0.12.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
xyzservices==2023.10.1
